chore: split stories by user
