+++ packages/client/src/v2-events/features/events/actions/validate/Review.interaction.localRegistrar.stories.tsx
@@ -24,7 +24,6 @@ import { ROUTES, routesConfig } from '@client/v2-events/routes'
 import { useEventFormData } from '@client/v2-events/features/events/useEventFormData'
 import { AppRouter } from '@client/v2-events/trpc'
 import { testDataGenerator } from '@client/tests/test-data-generators'
-import { tennisClubMembershipEventIndex } from '@client/v2-events/features/events/fixtures'
 import { createDeclarationTrpcMsw } from '../../../../../tests/v2-events/declaration.utils'
 import { Review } from './index'
 
@@ -36,7 +35,7 @@ const declareEventDocument = generateEventDocument({
 })
 
 const meta: Meta<typeof Review> = {
-  title: 'Validate/Interaction',
+  title: 'Validate/Review/Interaction/Local Registrar',
   beforeEach: () => {
     useEventFormData.setState({
       formValues: getCurrentEventState(declareEventDocument).declaration
@@ -56,33 +55,12 @@ const tRPCMsw = createTRPCMsw<AppRouter>({
   transformer: { input: superjson, output: superjson }
 })
 
-const callTracker = {
-  localRegistrar: {
-    'event.create': 0,
-    'event.actions.declare': 0,
-    'event.actions.validate': 0,
-    'event.actions.register': 0,
-    'event.actions.archive': 0,
-    'event.actions.reject': 0
-  },
-  registrationAgent: {
-    'event.create': 0,
-    'event.actions.declare': 0,
-    'event.actions.validate': 0,
-    'event.actions.register': 0,
-    'event.actions.archive': 0,
-    'event.actions.reject': 0
-  }
-}
-
 const eventDocument = generateEventDocument({
   configuration: tennisClubMembershipEvent,
   actions: [ActionType.CREATE]
 })
 
 const eventId = eventDocument.id
-
-const draft = generateEventDraftDocument(eventId, ActionType.REGISTER)
 
 const mockUser = {
   id: '67bda93bfc07dee78ae558cf',
@@ -165,76 +143,6 @@ export const ReviewForLocalRegistrarCompleteInteraction: Story = {
   }
 }
 
-export const ReviewForRegistrationAgentCompleteInteraction: Story = {
-  beforeEach: () => {
-    window.localStorage.setItem(
-      'opencrvs',
-      generator.user.token.registrationAgent
-    )
-    declarationTrpcMsw.events.reset()
-    declarationTrpcMsw.drafts.reset()
-  },
-  parameters: {
-    reactRouter: {
-      router: routesConfig,
-      initialPath: ROUTES.V2.EVENTS.VALIDATE.REVIEW.buildPath({
-        eventId: declareEventDocument.id
-      })
-    },
-    chromatic: { disableSnapshot: true },
-    msw: {
-      handlers: {
-        drafts: declarationTrpcMsw.drafts.handlers,
-        events: declarationTrpcMsw.events.handlers,
-        user: [
-          graphql.query('fetchUser', () => {
-            return HttpResponse.json({
-              data: {
-                getUser: generator.user.registrationAgent()
-              }
-            })
-          }),
-          tRPCMsw.user.list.query(([id]) => {
-            return [mockUser]
-          })
-        ]
-      }
-    }
-  },
-  play: async ({ canvasElement, step }) => {
-    await step('Modal has scope based content', async () => {
-      const canvas = within(canvasElement)
-      await userEvent.click(
-        await canvas.findByRole('button', { name: 'Send for approval' })
-      )
-
-      const modal = within(await canvas.findByRole('dialog'))
-
-      await modal.findByText('Send for approval?')
-      await modal.findByRole('button', { name: 'Cancel' })
-      await userEvent.click(
-        await modal.findByRole('button', { name: 'Confirm' })
-      )
-    })
-
-    await step('Confirm action triggers scope based actions', async () => {
-      await within(canvasElement).findByText('All events')
-
-      await waitFor(async () => {
-        await expect(declarationTrpcMsw.events.getSpyCalls()).toMatchObject({
-          'event.create': false,
-          'event.actions.notify.request': false,
-          'event.actions.declare.request': false,
-          'event.actions.validate.request': true,
-          'event.actions.register.request': false,
-          'event.actions.archive.request': false,
-          'event.actions.reject.request': false
-        })
-      })
-    })
-  }
-}
-
 export const ReviewForLocalRegistrarArchiveInteraction: Story = {
   beforeEach: () => {
     window.localStorage.setItem('opencrvs', generator.user.token.localRegistrar)
@@ -251,11 +159,7 @@ export const ReviewForLocalRegistrarArchiveInteraction: Story = {
     chromatic: { disableSnapshot: true },
     msw: {
       handlers: {
-        drafts: [
-          tRPCMsw.event.draft.list.query(() => {
-            return [draft]
-          })
-        ],
+        drafts: declarationTrpcMsw.drafts.handlers,
         events: declarationTrpcMsw.events.handlers,
         user: [
           graphql.query('fetchUser', () => {
@@ -266,115 +170,6 @@ export const ReviewForLocalRegistrarArchiveInteraction: Story = {
             })
           }),
           tRPCMsw.user.list.query(() => {
-            return [mockUser]
-          })
-        ]
-      }
-    }
-  },
-  play: async ({ canvasElement, step }) => {
-    const canvas = within(canvasElement)
-
-    await step('Open reject modal', async () => {
-      const rejectButton = await canvas.findByRole('button', {
-        name: 'Reject'
-      })
-      await userEvent.click(rejectButton)
-    })
-
-    const modal = within(await canvas.findByRole('dialog'))
-
-    await step('Archive is disabled', async () => {
-      const archiveButton = await modal.findByRole('button', {
-        name: 'Archive'
-      })
-
-      await expect(archiveButton).toBeDisabled()
-    })
-
-    await step('Add description', async () => {
-      const descriptionInput = await modal.findByRole('textbox')
-
-      await fireEvent.change(descriptionInput, {
-        target: { value: 'Wrong data' }
-      })
-    })
-
-    await step('Archive is not disabled', async () => {
-      const archiveButton = await modal.findByRole('button', {
-        name: 'Archive'
-      })
-
-      await expect(archiveButton).not.toBeDisabled()
-    })
-
-    await step('Mark as a duplicate', async () => {
-      const markAsDuplicateCheckbox = await modal.findByRole('checkbox', {
-        name: 'Mark as a duplicate'
-      })
-
-      await userEvent.click(markAsDuplicateCheckbox)
-    })
-
-    await step('Archive is not disabled', async () => {
-      const archiveButton = await modal.findByRole('button', {
-        name: 'Archive'
-      })
-
-      await expect(archiveButton).not.toBeDisabled()
-
-      await userEvent.click(archiveButton)
-      await within(canvasElement).findByText('All events')
-
-      await waitFor(async () => {
-        await expect(declarationTrpcMsw.events.getSpyCalls()).toMatchObject({
-          'event.create': false,
-          'event.actions.notify.request': false,
-          'event.actions.declare.request': false,
-          'event.actions.validate.request': false,
-          'event.actions.register.request': false,
-          'event.actions.archive.request': true,
-          'event.actions.reject.request': false
-        })
-      })
-    })
-  }
-}
-
-export const ReviewForRegistrationAgentArchiveInteraction: Story = {
-  beforeEach: () => {
-    window.localStorage.setItem(
-      'opencrvs',
-      generator.user.token.registrationAgent
-    )
-    declarationTrpcMsw.events.reset()
-    declarationTrpcMsw.drafts.reset()
-  },
-  parameters: {
-    reactRouter: {
-      router: routesConfig,
-      initialPath: ROUTES.V2.EVENTS.VALIDATE.REVIEW.buildPath({
-        eventId: declareEventDocument.id
-      })
-    },
-    chromatic: { disableSnapshot: true },
-    msw: {
-      handlers: {
-        drafts: [
-          tRPCMsw.event.draft.list.query(() => {
-            return [draft]
-          })
-        ],
-        events: declarationTrpcMsw.events.handlers,
-        user: [
-          graphql.query('fetchUser', () => {
-            return HttpResponse.json({
-              data: {
-                getUser: generator.user.registrationAgent()
-              }
-            })
-          }),
-          tRPCMsw.user.list.query(([id]) => {
             return [mockUser]
           })
         ]
